+++ react/components/FixedButton/index.js
@@ -1,12 +1,12 @@
 import React, { Fragment } from 'react'
 import BottomExtraSpace from './BottomExtraSpace'
-import { NoSSR } from 'render'
+import NoSsr from '@material-ui/core/NoSsr'
 
 export default props => (
   <Fragment>
     <div className="fixed bottom-0 left-0 w-100 z-999">{props.children}</div>
-    <NoSSR>
+    <NoSsr>
       <BottomExtraSpace>{props.children}</BottomExtraSpace>
-    </NoSSR>
+    </NoSsr>
   </Fragment>
 )